Removed files of previous workflow
It turned out this is not the workflow we want to be streamlining/automating here. Files associated with it will be archived on the JSDM_workflow_shelved in case we need them again later.
The main branch will be used for streamlining the workflows that are routinely used by Diego et al.