Adapted MSI function to run in workflow
This included a range of smaller changes to both the data preparation part and the core code originating from PECBSM.
Changed over to using the "COMB_[....]_indices_TT" files as input instead of the ones from Hekkefuglovervåkingen only (but note that this should be made flexible).
Had to remove the "as.vector()" calls in the PECBMS code as they resulted in the creation of list objects that then threw errors when subjected to "log()" further down.
Also added an argument to specify a (new) directory for storing the results.