Fixed issue with broken indexing in RSWAN::Rtrimcovin
There was a row selection via index in the function (original PECBMS files) that no longer worked because for some reason, the 1 in the data frame has non-0 digits from digit 16 onwards. This made the indexing condition not recognize it, returning an empty dataframe.

NOTE: This is a roundabout fix in the original file from PECBMS. It is not fixed in the newer version. So we will have to fix it again when we update the script to the newer version.